refactor: 테스트 통합화
